+++ src/Raven.CodeAnalysis/Macros/DefaultMacroEnvironment.cs
@@ -5,5 +5,5 @@ namespace Raven.CodeAnalysis.Macros;
 internal static class DefaultMacroEnvironment
 {
     public static ImmutableArray<IMacroDefinition> Macros { get; } =
-        [IntrinsicQuoteMacro.Instance];
+        [IntrinsicQuoteMacro.Instance, IntrinsicCompileMacro.Instance];
 }